# perf: rule runner