--- a/packages/cli/src/unminify.worker.ts
+++ b/packages/cli/src/unminify.worker.ts
@@ -1,12 +1,9 @@
 /* eslint-disable no-console */
-import path from 'node:path'
-import process from 'node:process'
 import { Timing } from '@wakaru/ast-utils/timing'
-import { runTransformations, transformationRules } from '@wakaru/unminify'
+import { runDefaultTransformation } from '@wakaru/unminify'
 import fsa from 'fs-extra'
 import { ThreadWorker } from 'poolifier'
 import type { UnminifyWorkerParams } from './types'
-import type { Transform } from 'jscodeshift'
 
 export async function unminify(data?: UnminifyWorkerParams) {
     if (!data) throw new Error('No data received')
@@ -14,24 +11,10 @@ export async function unminify(data?: UnminifyWorkerParams) {
     const timing = new Timing()
     const { inputPath, outputPath, moduleMeta, moduleMapping } = data
     try {
-        const cwd = process.cwd()
-        const filename = path.relative(cwd, inputPath)
         const source = await fsa.readFile(inputPath, 'utf-8')
         const fileInfo = { path: inputPath, source }
 
-        const transformations = transformationRules.map<Transform>((rule) => {
-            const { id, transform } = rule
-            const fn = (...args: Parameters<Transform>) => {
-                const stopMeasure = timing.startMeasure(filename, id)
-                transform(...args)
-                stopMeasure()
-            }
-            // Set the name of the function for better debugging
-            Object.defineProperty(fn, 'name', { value: id })
-            return fn
-        })
-
-        const { code } = runTransformations(fileInfo, transformations, { moduleMeta, moduleMapping })
+        const { code } = runDefaultTransformation(fileInfo, { moduleMeta, moduleMapping })
         await fsa.ensureFile(outputPath)
         await fsa.writeFile(outputPath, code, 'utf-8')