--- a/packages/cli/src/unminify.worker.ts
+++ b/packages/cli/src/unminify.worker.ts
@@ -1,20 +1,19 @@
 /* eslint-disable no-console */
-import { Timing } from '@wakaru/shared/timing'
-import { runDefaultTransformation } from '@wakaru/unminify'
+import { runDefaultTransformationRules } from '@wakaru/unminify'
 import fsa from 'fs-extra'
 import { ThreadWorker } from 'poolifier'
 import type { UnminifyWorkerParams } from './types'
+import type { Timing } from '@wakaru/shared/timing'
 
 export async function unminify(data?: UnminifyWorkerParams) {
     if (!data) throw new Error('No data received')
 
-    const timing = new Timing()
     const { inputPath, outputPath, moduleMeta, moduleMapping } = data
     try {
         const source = await fsa.readFile(inputPath, 'utf-8')
         const fileInfo = { path: inputPath, source }
 
-        const { code } = runDefaultTransformation(fileInfo, { moduleMeta, moduleMapping })
+        const { code, timing } = runDefaultTransformationRules(fileInfo, { moduleMeta, moduleMapping })
         await fsa.ensureFile(outputPath)
         await fsa.writeFile(outputPath, code, 'utf-8')
 
@@ -25,8 +24,8 @@ export async function unminify(data?: UnminifyWorkerParams) {
         console.log()
         console.error(e)
 
-        return timing
+        return null
     }
 }
 
-export default new ThreadWorker<UnminifyWorkerParams, Timing>(unminify)
+export default new ThreadWorker<UnminifyWorkerParams, Timing | null>(unminify)