This PR changed how a rule was created and provide an improved way to run the rules.

Tested with 112 modules (1MB)

| Concurrency | Before | After | Diff |
| ----------- | ------ | ----- | ---- |
| 1           | 181s   | 71s   | **2.5x** |
| 12          | 132s   | 50s   | **2.6x** |

`jscodeshift` actually spends a lot of time on parsing.

```
┌─────────┬──────────────────────────┬───────┐
│ (index) │           key            │ time  │
├─────────┼──────────────────────────┼───────┤
│    0    │       'un-boolean'       │  279  │
│    1    │     'un-while-loop'      │  290  │
│    2    │  'un-flip-comparisons'   │  294  │
│    3    │       'un-return'        │  307  │
│    4    │  'un-template-literal'   │  343  │
│    5    │     'un-use-strict'      │  369  │
│    6    │  'un-bracket-notation'   │  381  │
│    7    │   'un-numeric-literal'   │  391  │
│    8    │      'un-undefined'      │  392  │
│    9    │ 'un-assignment-merging'  │  462  │
│   10    │  'un-variable-merging'   │  532  │
│   11    │        'un-iife'         │  563  │
│   12    │       'un-typeof'        │  567  │
│   13    │     'un-async-await'     │  606  │
│   14    │      'un-infinity'       │  659  │
│   15    │    'un-esmodule-flag'    │  660  │
│   16    │  'un-type-constructor'   │  835  │
│   17    │    'un-conditionals'     │  946  │
│   18    │     'module-mapping'     │  962  │
│   19    │        'un-enum'         │  973  │
│   20    │      'un-es6-class'      │ 1185  │
│   21    │     'un-parameters'      │ 1506  │
│   22    │         'un-jsx'         │ 1551  │
│   23    │  'un-builtin-prototype'  │ 1744  │
│   24    │    'un-export-rename'    │ 1980  │
│   25    │ 'un-nullish-coalescing'  │ 3001  │
│   26    │  'un-optional-chaining'  │ 3136  │
│   27    │       'prettier-1'       │ 3495  │
│   28    │    'un-curly-braces'     │ 3682  │
│   29    │      'smart-rename'      │ 3851  │
│   30    │      'smart-inline'      │ 4222  │
│   31    │         'un-esm'         │ 4543  │
│   32    │   'jscodeshift-print'    │ 5464  │
│   33    │        'prettier'        │ 6318  │
│   34    │  'un-parameters-lebab'   │ 6888  │
│   35    │   'un-runtime-helper'    │ 9947  │
│   36    │ 'un-sequence-expression' │ 11609 │
│   37    │   'jscodeshift-parse'    │ 12072 │
│   38    │         'lebab'          │ 13596 │
│   39    │    'un-indirect-call'    │ 15940 │
└─────────┴──────────────────────────┴───────┘
```