@pionxzh 2024-01-05
I'm away from my pc. But this PR can bring down the total execution time of a 1.5 MB sample from 2.x minutes to ~50s 🎉